requirements.txt
PyQt5==5.10.1
PyYAML==3.13
sip==4.19.8
